x1plusd: refactor to split DBus logic into its own class, which SettingsService subclasses (prepare for x1plusd-ota module)
