+++ images/cfw/opt/x1plus/lib/python/x1plus/utils.py
@@ -1,0 +1,18 @@
+import subprocess
+from functools import lru_cache
+import os
+
+
+@lru_cache(None)
+def is_emulating():
+    return not os.path.exists("/etc/bblap")
+
+
+@lru_cache(None)
+def serial_number():
+    """
+    Used to get the Serial Number for the Printer
+    """
+    return subprocess.check_output(["bbl_3dpsn"], stderr=subprocess.DEVNULL).decode(
+        "utf-8"
+    )
